% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_query_functions.R
\name{create_elvis_query_url}
\alias{create_elvis_query_url}
\title{Create an ELVIS url for a given bounding box}
\usage{
create_elvis_query_url(
  x,
 
    base_url = "https://elvis2018-ga.fmecloud.com/fmedatastreaming/elvis_indexes/ReturnDownloadables.fmw"
)
}
\arguments{
\item{x}{An object from which to derive a bounding box for the query. Can be
any of the following: 
\itemize{
  \item{An \code{sf} bounding box object (e.g. created using the 
    \code{st_bbox} function).}
  \item{An \code{sf} object from which a non-empty bounding box can be 
    derived, e.g. a spatial data frame, or a single polygon or line 
    feature, but not a point geometry.} 
  \item{A named, four-element numeric vector where the names are 
    \code{'xmin', 'ymin', 'xmax', 'ymax'}. }
  \item{An unnamed four-element numeric vector, assumed to be ordered as
    \code{xmin, ymin, xmax, ymax}. }
 }}

\item{base_url}{The base URL for the ELVIS archive. Currently this is:
\code{"https://elvis2018-ga.fmecloud.com/fmedatastreaming/elvis_indexes/ReturnDownloadables.fmw"}.}
}
\value{
A URL to query ELVIS.
}
\description{
This function is primarily intended as a helper that is called by the
\code{\link{query_elvis}} function, but you might want to use it directly,
e.g. to create a URL that you then paste into a browser.
}
